alphabrain/pyproject.toml
[tool.poetry]
name = "ghostquant-alphabrain"
version = "0.1.0"
description = "GhostQuant AlphaBrain Quant Intelligence Module"
authors = ["GhostQuant Team"]

[tool.poetry.dependencies]
python = "^3.11"
psycopg = {extras = ["binary", "pool"], version = "^3.1.0"}
redis = "^5.0.0"
numpy = "^1.26.0"
pandas = "^2.1.0"
scikit-learn = "^1.3.0"
scipy = "^1.11.0"
aiohttp = "^3.9.0"
python-dotenv = "^1.0.0"
reportlab = "^4.0.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
